LayoutTests/editing/inserting/rebalance-whitespace-1-expected.txt

This tests whitespace rebalancing when inserting characters in and around already balanced sequences. In each div below, 2 or more adjacent whitespace characters should be alternate between regular whitespace and non-breaking spaces (note that editable regions use -webkit-nbsp-mode:space).

hellox world
hellox  world
hellox   world
hello x world
hello  x world
hello  x  world
hello  x   world
hello   x  world
hello   x   world
